♻️ refactor(rename): ThemeSwitcher now lowercase
