+++ frontend/components/theme-switcher/AuthTheme.tsx
